Homework_2/Space Scanner/src/bg/sofia/uni/fmi/mjt/space/algorithm/Rijndael.java
package bg.sofia.uni.fmi.mjt.space.algorithm;

import bg.sofia.uni.fmi.mjt.space.exception.CipherException;

import javax.crypto.Cipher;
import javax.crypto.SecretKey;
import java.io.InputStream;
import java.io.OutputStream;

public class Rijndael implements SymmetricBlockCipher {
    private static final String ENCRYPTION_ALGORITHM = "AES";
    private final SecretKey secretKey;

    public Rijndael(SecretKey secretKey) {
        this.secretKey = secretKey;
    }

    @Override
    public void encrypt(InputStream inputStream, OutputStream outputStream) throws CipherException {
        try {
            Cipher cipher = Cipher.getInstance(ENCRYPTION_ALGORITHM);
            cipher.init(Cipher.ENCRYPT_MODE, secretKey);
            outputStream.write(cipher.doFinal(inputStream.readAllBytes()));
        } catch (Exception e) {
            throw new CipherException("Something went wrong trying to encrypt the stream", e);
        }
    }

    @Override
    public void decrypt(InputStream inputStream, OutputStream outputStream) throws CipherException {
        try {
            Cipher cipher = Cipher.getInstance(ENCRYPTION_ALGORITHM);
            cipher.init(Cipher.DECRYPT_MODE, secretKey);
            outputStream.write(cipher.doFinal(inputStream.readAllBytes()));
        } catch (Exception e) {
            throw new CipherException("Something went wrong trying to decrypt the stream", e);
        }
    }
}
